bcm470x:update kernel

+++ release/linux/bcm/linux-4.4/include/net/secure_seq.h
@@ -2,10 +2,9 @@
 #define _NET_SECURE_SEQ
 
 #include <linux/types.h>
-#include <linux/in6.h>
 
-u64 secure_ipv4_port_ephemeral(__be32 saddr, __be32 daddr, __be16 dport);
-u64 secure_ipv6_port_ephemeral(const __be32 *saddr, const __be32 *daddr,
+u32 secure_ipv4_port_ephemeral(__be32 saddr, __be32 daddr, __be16 dport);
+u32 secure_ipv6_port_ephemeral(const __be32 *saddr, const __be32 *daddr,
 			       __be16 dport);
 __u32 secure_tcp_sequence_number(__be32 saddr, __be32 daddr,
 				 __be16 sport, __be16 dport);
